+++ app/assets/javascripts/application.js
@@ -5,6 +5,7 @@
 
 // APP SETUP
 //= require globals/base
+//= require helpers/url
 
 // CELLS
 //= require compare_provider_button
@@ -14,6 +15,7 @@
 //= require provider_list_half_right
 //= require provider_list_large
 //= require reference_map
+//= require site_header
 //= require story_slider
 
 // VISUALIZATION CELLS