+++ app/controllers/providers_controller.rb
@@ -1,3 +1,8 @@
 class ProvidersController < ApplicationController
   autocomplete :provider, :search_name, full: true, limit: 15
+
+  def index
+    @providers = Provider.where("name ILIKE ?", "%#{params[:q]}%")
+    render json: @providers
+  end
 end