+++ CHANGELOG.md
@@ -5,11 +5,34 @@ All notable changes to this project will be documented in this file.
 ## 2018-07-05
 
 ### Modified
+  - `app/assets/images/icons/menu.svg`
+  - `app/assets/javascripts/application.js`
+  - `app/assets/javascripts/globals/atsb.js`
+  - `app/assets/javascripts/globals/base.js`
+  - `app/assets/stylesheets/layout/_application.scss`
+  - `app/cells/components/compare_provider_button/compare_provider_button.scss`
+  - `app/cells/components/provider_detail_half_left/provider_detail_half_left.scss`
+  - `app/cells/components/provider_detail_half_right/provider_detail_half_right.scss`
+  - `app/cells/components/provider_detail_large/provider_detail_large.scss`
+  - `app/cells/components/provider_list_half_right/provider_list_half_right.scss`
+  - `app/cells/components/provider_list_large/provider_list_large.js`
+  - `app/cells/components/provider_list_large/provider_list_large.scss`
+  - `app/cells/components/provider_list_large/show.slim`
   - `app/cells/components/site_header/show.slim`
   - `app/cells/components/site_header/site_header.scss`
+  - `app/cells/components/story_slider/show.slim`
+  - `app/cells/components/story_slider/story_slider.scss`
+  - `app/controllers/providers_controller.rb`
+  - `app/views/home/index.slim`
+  - `app/views/layouts/public.slim`
+  - `config/application.rb`
+  - `config/routes.rb`
 
 ### Added
   - `app/assets/images/logos/atsb-w.svg`
+  - `app/assets/javascripts/globals/fetch.js`
+  - `app/assets/javascripts/helpers/url.js`
+  - `app/cells/components/site_header/site_header.js`
 
 ## 2018-07-04